stack: MpiWcsSelectImagesTask is not necessary
With the improved MpiDataIdContainer (in MpiArgumentParser), there's no
need to wait to receive the selection data.  This was the cause of a
general slow-down (because everything had to keep waiting for the
broadcast) and deadlock (because other nodes could finish their pile
and move on, so that not everyone was on the same page).