detrends: mask NANs in combined image
We may have a NAN in the combined image if there were no unmasked
inputs to a particular pixel.  The result should have a mask value
of NO_DATA, but the NAN needs to be suppressed so it doesn't
propagate into the image when we do ISR with this detrend.