detrends: make NAN masking more generic
Adds support for masking NANs in an Exposure and DecoratedImage.
Changed the style of masking to simply replace NANs with the
median using numpy rather than interpolating, as the interpolation
code is specific to an Exposure (it ideally needs a Ccd so it
can tell whether the rows are vertical or horizontal).