ProcessExposure: allow CCDs to be ignored
The "ignoreCcdList" functionality was removed when the focus
processing was added (since it was a mechanism to ignore the
focus CCDs, but they were now supported).  However, this
functionality is useful for restricting the amount of processing
required in testing.

Added an 'ignoreCcdList' parameter to the ProcessExposureConfig,
allowing individual CCDs (specified by identifier) to be ignored
and some small changes to support this functionality.