ProcessExposureTask: add global curve of growth calculation
Using a separate curve of growth per CCD can lead to discontinuities
across the focal plane (e.g., if one CCD can't correct to the same
radius as all the others), and these can't be fixed by mosaicking (as
that allows exposure offsets and CCD offsets, but not exposure+CCD
offsets). Therefore we gather the calibration catalogs and use that
to create a curve of growth for the entire exposure that gets exported
to the slave nodes and applied to the final catalogs.