multiband: implement reprocessing strategy for dealing with large footprints
This exists as a workaround for large deblender footprints causing large
memory use or very slow processing.  We refuse to deblend those footprints
when running on a cluster and return to "reprocess" on a machine with
larger memory or more time if we consider those footprints important to
recover.

If we're reprocessing (add, e.g., "-c reprocessing=True
measureCoaddSources.deblend.maxFootprintArea=XXX --clobber-config" to
command-line), then we should breeze through the detection and detection
merge steps, then we discover which patches require reprocessing and touch
a file on disk to indicate we're in the process of reprocessing the patch
(in case the reprocessing is interrupted, it would leave things in an
inconsistent state because it clobbers existing outputs; the file allows
us to trigger reprocessing on patches that have already been
half-reprocessed) and proceed with the remaining steps.

One drawback of this scheme as implemented is that we're essentially
flushing away the configuration information (we're re-running the
processing with a different deblender maxFootprintArea config, which
requires the --clobber-config argument) which may lead to confusion down
the road.  Also, the entire patch is reprocessed instead of only the
large footprints; this simplifies the code.